requirements.txt
pandas>=1.5.0
openpyxl>=3.0.0
xlrd>=2.0.0
matplotlib>=3.5.0
seaborn>=0.11.0
requests>=2.28.0
numpy>=1.21.0
psycopg2-binary